Attempted to refine matches
Added assumption that DataFrame object will match pattern of ".*df.*"
which ruled out a lot of false positives.

Some repos seem to generate a lot of false positives during ipynb to py
conversions. More specifically, it seems to match substrings in a "true"
match but the matches list usually contains the correct parts.